Commit 12.06

+++ app/src/Controller/CategoryController.php
@@ -169,6 +169,11 @@ class CategoryController extends AbstractController
      */
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if ($category->getPosts()->count()) {
+            $this->addFlash('warning', 'message_category_contains_posts');
+
+            return $this->redirectToRoute('category_index');
+        }
         $form = $this->createForm(CategoryType::class, $category, ['method' => 'DELETE']);
         $form->handleRequest($request);
